Add vertex array support

+++ Wizar/src/Wizar/ImGui/ImGuiLayer.cpp
@@ -74,7 +74,7 @@ namespace Wizar {
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().GetWidth(),app.GetWindow().GetHeight());
+		io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(),(float)app.GetWindow().GetHeight());
 
 		// Rendering
 		ImGui::Render();
@@ -95,7 +95,5 @@ namespace Wizar {
 	{
 		static bool show = true;
 		ImGui::ShowDemoWindow(&show);
-		
-
 	}
 }
